Move worker to object manager
