+++ src/device.rs
@@ -9,57 +9,12 @@ use crate::instance::InstanceContext;
 use crate::util::extensions::{AsRefOption, ExtensionFunctionSet, VkExtensionInfo, VkExtensionFunctions};
 use crate::UUID;
 
-struct WorkerData {
-    queue: ConcurrentQueue<()>,
-}
-
-impl WorkerData {
-    fn new() -> Self {
-        Self {
-            queue: ConcurrentQueue::unbounded(),
-        }
-    }
-
-    /// After this function is called no new requests may be pushed into the worker
-    pub fn close(&self) {
-        self.queue.close();
-    }
-}
-
-struct Worker {
-    thread: Option<JoinHandle<()>>,
-    data: Arc<WorkerData>,
-}
-
-impl Worker {
-    pub fn spawn() -> Self {
-        let data = Arc::new(WorkerData::new());
-        let cloned_data = data.clone();
-        Self {
-            thread: Some(std::thread::spawn(move|| Self::run(cloned_data))),
-            data,
-        }
-    }
-
-    fn run(data: Arc<WorkerData>) {
-
-    }
-}
-
-impl Drop for Worker {
-    fn drop(&mut self) {
-        self.data.close();
-        self.thread.take().unwrap().join().unwrap();
-    }
-}
-
 pub struct DeviceContextImpl {
     instance: InstanceContext,
     device: ash::Device,
     physical_device: vk::PhysicalDevice,
     extensions: ExtensionFunctionSet,
     features: EnabledFeatures,
-    worker: Worker,
 }
 
 impl Drop for DeviceContextImpl {
@@ -81,7 +36,6 @@ impl DeviceContext {
             physical_device,
             extensions,
             features,
-            worker: Worker::spawn(),
         }))
     }
